requirements.txt
alembic==1.11.1
APIFlask==1.3.1
boto3==1.26.114
botocore==1.29.147
firebase_admin==6.2.0
Flask==2.2.1
Flask_Cors==3.0.10
Flask_Migrate==4.0.4
Flask_SocketIO==5.3.4
flask_sqlalchemy==3.0.3
openai==0.27.8
Pillow==9.5.0
PyJWT==2.7.0
python-dotenv==1.0.0
python_dateutil==2.8.2
Requests==2.31.0
requests_aws4auth==1.2.3
SQLAlchemy==2.0.15
sqlalchemy_easy_softdelete==0.8.2
stripe==5.4.0
tiktoken==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
elevenlabs==0.2.24
psycopg2-binary==2.9.7
